+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * db: add experimental postgresql support
   > many new db products use postgres compatible driver, e.g. GCloud AlloyDB, CockroachDB
   > PostgreSQL lacks of many features we are using with MySQL, 1. affected rows, 2. QueryInterceptor to track no index used sql
+* search: clear ForEach scroll once process done
 
 ### 8.0.3 (06/22/2022 - 07/08/2022)